Moves repositories
